+++ src/main/python/backend/saver/ICSessionHandler.py
@@ -21,10 +21,12 @@ class ICSessionHandler(object):
             return False, "No data loaded."
 
         #get main figures
-        mainFigures = self.mC.mainFrames["right"].mainFigureRegistry.getMainFigures()
+        mainFigures = self.mC.mainFrames["right"].mainFigureRegistry.getMainFiguresByID()
         print(mainFigures)
-
-        combinedObj = {"dfs":dataFrames,"dfsName":dataFrameNames,"mainFigures":mainFigures}
+        mainFigureRegistry = self.mC.mainFrames["right"].mainFigureRegistry
+        comboSettings = []#self.mC.mainFrames["right"].mainFigureRegistry.getMainFigureCurrentSettings() 
+        print(mainFigureRegistry.mainFigureTemplates)
+        combinedObj = {"dfs":dataFrames,"dfsName":dataFrameNames,"mainFigures":mainFigures,"mainFigureRegistry":mainFigureRegistry,"mainFigureComboSettings":comboSettings}
         print(combinedObj)
         with open(sessionPath,"wb") as icSession:
             pickle.dump(combinedObj,icSession)
@@ -42,4 +44,6 @@ class ICSessionHandler(object):
         print(response)
         #open mainf figures
         response["mainFigures"] = combinedObj["mainFigures"] #cannot be done from WorkingThread
+        response["mainFigureRegistry"] = combinedObj["mainFigureRegistry"]
+        response["mainFigureComboSettings"] = []#combinedObj["mainFigureComboSettings"]
         return response
